Add details to plant card styles
