+++ components/PlantCard/PlantCard.module.scss
@@ -4,11 +4,11 @@
   position: relative;
   perspective: 150rem;
   -moz-perspective: 150rem;
-  height: 28.5rem;
+  height: 30.5rem;
   border-radius: $borderRadius;
 
   div {
-    height: 28.5rem;
+    height: 30.5rem;
     transition: all 1s ease;
     position: absolute;
     top: 0;
