+++ lib/apolloClient/link.ts
@@ -1,29 +1,8 @@
-import { ApolloLink, HttpLink } from '@apollo/client';
-import { Auth } from 'aws-amplify';
+import { HttpLink } from '@apollo/client';
 
 const httpLink = new HttpLink({
   uri: 'http://localhost:3000/dev/graphql',
 });
 
-const getSession = async () => {
-  let session;
-  try {
-    session = await Auth.currentSession();
-    console.log(session);
-    return session;
-  } catch (e) {
-    console.log(e);
-  }
-};
-
-const AuthLink = new ApolloLink((operation, forward) => {
-  getSession();
-  operation.setContext(({ headers = {} }) => ({
-    headers: {
-      ...headers,
-    },
-  }));
-  return forward(operation);
-});
-
-export default AuthLink.concat(httpLink);
+export default httpLink;
+// 'https://mz4y6djivh.execute-api.us-east-1.amazonaws.com/dev/graphql'
