+++ serverless.yml
@@ -1,2 +1,6 @@
 blueSprucePlantGuide:
   component: "@sls-next/serverless-component@1.18.0"
+  inputs:
+    build:
+      env:
+        APP_URL:"https://mz4y6djivh.execute-api.us-east-1.amazonaws.com/dev"
